# docs/openapi.yaml
openapi: 3.0.3
info:
  title: YeYing SIWE Auth API
  version: 1.0.1
  description: |
    Backend API contract for the YeYing injected wallet SDK.

    Notes:
    - All responses are wrapped in a unified envelope: { code, message, data, timestamp }.
    - code = 0 indicates success; non-zero indicates failure and data should be null.
    - /refresh relies on the httpOnly refresh_token cookie; clients should send credentials.
servers:
  - url: http://localhost:3203
paths:
  /api/v1/public/auth/challenge:
    post:
      summary: Create login challenge
      description: Generate a SIWE challenge to be signed by the wallet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeRequest'
            examples:
              example:
                value:
                  address: "0xabc123..."
      responses:
        '200':
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'
        '400':
          description: Missing address or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/public/auth/verify:
    post:
      summary: Verify signature
      description: Verify the signed challenge and issue access + refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
            examples:
              example:
                value:
                  address: "0xabc123..."
                  signature: "0x..."
      responses:
        '200':
          description: Verified, access token returned and refresh cookie set
          headers:
            Set-Cookie:
              description: httpOnly refresh_token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Missing address/signature or challenge expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/public/auth/refresh:
    post:
      summary: Refresh access token
      description: Issue a new access token using the refresh_token cookie.
      responses:
        '200':
          description: Refreshed, access token returned and refresh cookie rotated
          headers:
            Set-Cookie:
              description: httpOnly refresh_token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Missing or invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/private/profile:
    get:
      summary: Get profile
      description: 'Example protected endpoint. Requires Authorization: Bearer <access-token> or <UCAN>.'
      security:
        - bearerAuth: []
        - ucanAuth: []
      responses:
        '200':
          description: Profile returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ucanAuth:
      type: http
      scheme: bearer
      bearerFormat: UCAN
  schemas:
    BaseResponse:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: integer
          description: 0 for success, non-zero for failure
          example: 0
        message:
          type: string
          example: ok
        timestamp:
          type: integer
          format: int64
          example: 1730000000000
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              nullable: true
              description: Always null when code != 0
    ChallengeRequest:
      type: object
      required:
        - address
      properties:
        address:
          type: string
          description: Wallet address (hex)
          example: 0xabc123...
    VerifyRequest:
      type: object
      required:
        - address
        - signature
      properties:
        address:
          type: string
          description: Wallet address (hex)
          example: 0xabc123...
        signature:
          type: string
          description: Signed challenge
          example: 0x...
    ChallengeData:
      type: object
      required:
        - address
        - challenge
        - nonce
        - issuedAt
        - expiresAt
      properties:
        address:
          type: string
        challenge:
          type: string
        nonce:
          type: string
        issuedAt:
          type: integer
          format: int64
        expiresAt:
          type: integer
          format: int64
    TokenData:
      type: object
      required:
        - address
        - token
        - expiresAt
        - refreshExpiresAt
      properties:
        address:
          type: string
        token:
          type: string
          description: Access token (JWT)
        expiresAt:
          type: integer
          format: int64
        refreshExpiresAt:
          type: integer
          format: int64
...
    ProfileData:
      type: object
      required:
        - address
        - issuedAt
      properties:
        address:
          type: string
        issuedAt:
          type: integer
          format: int64
    ChallengeResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/ChallengeData'
    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/TokenData'
...
    ProfileResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - data
          properties:
            data:
              $ref: '#/components/schemas/ProfileData'